Solution to Find Divisibility by 3 or 5

+++ Find Divisibility by 3 or 5@@6788dd63490637cfc44e1ba0/code.c
@@ -2,8 +2,8 @@
 
 int main()
 {
-    float a;
-    scanf("%f",&a);
+    int a;
+    scanf("%d",&a);
     if(a%3==0){
         printf("Divisible by 3");
     }
